Uniquify the socket path within a given process
Also move the code a little. This allows two nspluginwrapper plugins
with the same file name to be loaded at once in the same process. This
is not very important, except it appears Arora or QtWebKit call
NP_Initialize when probing plugins (instead of skipping to
NP_GetMIMEDescription and NP_GetValue like everyone else). They also
leave the plugin around evidently.

The result is that, if you have two versions of a wrapper Flash, the
second's will fail to bind the socket and then hang later on.