Check for NULL plugin instance values in NPN_GetValue and NPN_SetValue
This is to work-around Chromium not doing the check itself. It's only
somewhat Chromium's fault as they're perfectly okay with segfaults
(separate plugin process) and a plugin shouldn't request
NPNVWindowNPObject on a NULL npp anyway. And, in fact, they don't. This
is a side effect of our RPC system not being able to distinguish a NULL
plugin instance from an invalid plugin instance.

This check should be removed when the fix in Chromium has trickled down
to the stable channel or when our RPC system becomes less awful.