Rework the sync mechanism
We still request a SYNC on new invokes, as before. However, we do not
unblock the SYNC with a SYNC_ACK (and correspondingly the SYNC_ACK with
a SYNC_END) until we return to the event loop. Originally, we unblock
when the current invoke is finished, so conflicts always alternate
method calls. This makes race conditions insane. Instead, we interleave
one event loop iteration's worth of invokes as a unit.

The end result should be that the wrapper never sees inconsistencies
(modulo main loop source priorities but those were already violated).
The viewer still may. However its inconsistencies are limited to the
first RPC'd NPAPI call in a single event loop iteration. We can get rid
of this by driving the viewer event loop ourselves and surrounding every
dispatch with a SYNC/SYNC_END pair (thus binding the two event loops
together), but this may be prohibitively expensive for dispatches that
don't call NPAPI.

Closes #14. The crash was caused by an NPN_Evaluate (triggering an
NPP_SetWindow browser-side) getting in between a
NPP_SetWindow/NPP_HandleEvent pair because windowless plugins on Linux
make no sense.