Release pointer grabs before forwarding a ButtonPress to the plugin
Otherwise we still carry an implicit passive grab and the viewer cannot
upgrade to an active grab to show a context menu. This really only fixes
older Firefoxes as browsers which push plugins out-of-process themselves
(as they should) need to also include this fix. In theory this would
also fix Arora, but they have a QtWebKit bug where the event time is
bogus.

See browser bugs
http://code.google.com/p/chromium/issues/detail?id=40157
http://bugzilla.mozilla.org/show_bug.cgi?id=544211

Closes #16.