Revert "Decrease the RPC message timeout from 30 seconds to 10"
This reverts commit 1b259a81191aeec65f4c59c8dc321b05bf052e27. If you
call a function from Flash that ends up opening an alert, the call
blocks until the user closes the alert, and the call eventually times
out. A 30 second timeout isn't much better, but is harder to trigger
accidentally, whereas it's not difficult to trigger a 10 second timeout
on accident.

See bug #24.